chore: little more refactoring
