+++ src/commands.ts
@@ -3,7 +3,7 @@ import { commands } from 'vscode'
 import { EXT_NAMESPACE } from './meta'
 import { pause, start, stop } from './actions'
 
-export function RegisterCommands(ctx: ExtensionContext): ExtensionContext {
+export function registerCommands(ctx: ExtensionContext): ExtensionContext {
   ctx.subscriptions.push(
     commands.registerCommand(`${EXT_NAMESPACE}.start-automate`, async () => {
       await start()
