+++ src/utils/index.ts
@@ -1,5 +1,3 @@
-export * from './vscode'
-export * from './types'
+export * from './fs'
 export * from './obj'
-
-// export * from './fs' // TODO: Need to work - can't be exported due to unit testing issue
+export * from './vscode'
